Extract the distinct colors code into a reusable function.
This required moving the error module up to the root so `distinct` could access it.
