+++ src/distinct.rs
@@ -1,8 +1,12 @@
 use core::f64 as scalar;
+use std::io::{self, Write};
+
 
 use rand::prelude::*;
 
 use crate::delta_e;
+use crate::ansi::{Brush,Stream,ToAnsiStyle};
+use crate::error::{PastelError,Result};
 use crate::random::{self, RandomizationStrategy};
 use crate::{Color, Lab};
 
@@ -170,6 +174,109 @@ impl SimulatedAnnealing {
         result
     }
 }
+
+fn print_iteration(out: &mut dyn Write, brush: &Brush, stats: &IterationStatistics) -> Result<()> {
+    let result = stats.distance_result;
+    write!(
+        out,
+        "[{:10.}] D_mean = {:<6.2}; D_min = {:<6.2}; T = {:.6} ",
+        stats.iteration,
+        result.mean_closest_distance,
+        result.min_closest_distance,
+        stats.temperature
+    )?;
+    print_colors(out, brush, &stats.colors, Some(result.closest_pair))?;
+    Ok(())
+}
+
+fn print_colors(
+    out: &mut dyn Write,
+    brush: &Brush,
+    colors: &[Color],
+    closest_pair: Option<(usize, usize)>,
+) -> Result<()> {
+    let mut ci = 0;
+    for c in colors.iter() {
+        let tc = c.text_color();
+        let mut style = tc.ansi_style();
+        style.on(c);
+
+        if let Some(pair) = closest_pair {
+            if pair.0 == ci || pair.1 == ci {
+                style.bold(true);
+                style.underline(true);
+            }
+        }
+
+        write!(
+            out,
+            "{} ",
+            brush.paint(format!("{}", c.to_rgb_hex_string(false)), style)
+        )?;
+
+        ci += 1;
+    }
+    writeln!(out, "")?;
+    Ok(())
+}
+
+pub fn distinct_colors(
+    count: usize,
+    distance_metric: DistanceMetric,
+    verbose: bool,
+    arrange_colors: bool
+) -> Result<(Vec<Color>, DistanceResult)> {
+
+    if count < 2 {
+        return Err(PastelError::DistinctColorCountMustBeLargerThanOne);
+    }
+
+    let stderr = io::stderr();
+    let brush_stderr = Brush::from_environment(Stream::Stderr);
+
+    let mut colors = Vec::new();
+    for _ in 0..count {
+        colors.push(random::strategies::UniformRGB.generate());
+    }
+
+    let mut annealing = SimulatedAnnealing::new(
+        &colors,
+        SimulationParameters {
+            initial_temperature: 3.0,
+            cooling_rate: 0.95,
+            num_iterations: 100_000,
+            opt_target: OptimizationTarget::Mean,
+            opt_mode: OptimizationMode::Global,
+            distance_metric,
+        },
+    );
+
+    let mut callback: Box<dyn FnMut(&IterationStatistics)> = if verbose {
+        Box::new(|stats: &IterationStatistics| {
+            print_iteration(&mut stderr.lock(), &brush_stderr, stats).ok();
+        })
+    } else {
+        Box::new(|_: &IterationStatistics| {})
+    };
+
+    annealing.run(callback.as_mut());
+
+    annealing.parameters.initial_temperature = 0.5;
+    annealing.parameters.cooling_rate = 0.98;
+    annealing.parameters.num_iterations = 200_000;
+    annealing.parameters.opt_target = OptimizationTarget::Min;
+    annealing.parameters.opt_mode = OptimizationMode::Local;
+
+    let result = annealing.run(callback.as_mut());
+
+    let mut colors = annealing.get_colors();
+    if arrange_colors {
+        rearrange_sequence(&mut colors, distance_metric);
+    }
+
+    Ok((colors, result))
+}
+
 
 /// Re-arrange the sequence of colors such that the minimal difference between a given color and
 /// any of its predecessors is maximized.
